refactor: update user_id to int64 and standardize logging in auth-service
- Migrasi tipe data ID dari int32/string menjadi int64 pada auth.proto dan gRPC handler untuk konsistensi ekosistem microservices.
- Standarisasi format log menggunakan slog.Group pada middleware HTTP dan gRPC:
  * Group 'trace': Berisi request_id dan identitas user (user_id, username, role).
  * Group 'http'/'grpc': Berisi metadata protokol, status, dan latency dalam milidetik (duration_ms).
- Refactor helper log (CreateTraceGroup & DurationToMs) ke dalam middleware logger agar DRY (Don't Repeat Yourself).
- Cleanup inisialisasi gRPC server di main.go untuk memperbaiki double assignment server instance.
- Update gRPC handler GetUserProfile agar menggunakan parameter ID int64 secara langsung dari request.

+++ internal/entity/user.go
@@ -3,7 +3,7 @@ package entity
 import "time"
 
 type User struct {
-	ID           int       `gorm:"primaryKey;autoIncrement" json:"id"`
+	ID           int64     `gorm:"primaryKey;autoIncrement" json:"id"`
 	Username     string    `gorm:"unique;not null;type:varchar(50)" json:"username"`
 	PasswordHash string    `gorm:"not null;type:varchar(255)" json:"-"`
 	Email        string    `gorm:"unique;not null;type:varchar(100)" json:"email"`
